REFACTOR-#3539: align read_fwf with read_csv (#3540)
